requirements.txt
accelerate
altair
diffusers
hf-transfer
huggingface_hub
llama-index
llama-index-embeddings-huggingface
llama-index-llms-huggingface
lpips
moviepy
opencv-python
pandas
Pillow
pytorch-fid
scikit-learn
spacy
streamlit
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm
transformers
